pyproject.toml
[tool.poetry]
name = "geneset_translator"
version = "0.1.0"
description = "Explore biomedical knowledge graphs for your gene sets via NCATS Translator"
authors = ["Natalie Gill"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
tct = "^0.1.4"
streamlit = "^1.50.0"
pydantic = "^2.11.9"
networkx = "^3.5"
pandas = "^2.2.1"
python-louvain = "^0.16"
streamlit-cytoscape = ">=0.1.4"
scipy = "^1.16.2"
pyyaml = "^6.0.3"
treelib = "^1.8.0"
altair = "^6.0.0"
# Phase 3: LLM Summarization (optional)
anthropic = "^0.75.0"
python-dotenv = "^1.0.0"
requests = "^2.32.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
jupyter = "^1.1.1"
ipykernel = "^6.30.1"
watchdog = "^6.0.0"
notebook = "^7.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
